+++ CHANGELOG.md
@@ -3,6 +3,11 @@ CHANGELOG
 
 Notable changes will be documented in this file. The project adheres to [Semantic Versioning].
 
+Unreleased
+----------
+
+* Introduced `strict` and `fallback` mode (breaking change)
+
 7.1.0
 -----